requirements.txt
langchain_community
langchain-astradb
langchain-google-genai
datasets
pypdf
python-dotenv
flask
boto3

-e .